requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.4
flake8==6.0.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
mccabe==0.7.0
Pillow==10.0.0
pycodestyle==2.10.0
pydocstyle==6.3.0
pyflakes==3.0.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
